@@ -16,7 +16,7 @@
   "license": "MIT",
   "type": "module",
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=13.5.0"
   },
   "scripts": {
     "start": "cross-env DEBUG=*,-engine*,-babel*,-send,-ref*,-express* node -r dotenv/config --es-module-specifier-resolution=node index.js",
@@ -31,7 +31,7 @@
   "dependencies": {
     "@aquarius/regex": "^1.0.0",
     "@google-cloud/firestore": "^3.4.1",
-    "@sentry/node": "5.11.1",
+    "@sentry/node": "5.11.2",
     "alphavantage": "^2.0.0",
     "chalk": "^3.0.0",
     "cors": "^2.8.5",
@@ -52,7 +52,7 @@
     "pluralize": "^8.0.0",
     "pretty-bytes": "^5.0.0",
     "rgb-hex": "^3.0.0",
-    "rss-parser": "^3.7.3",
+    "rss-parser": "^3.7.4",
     "timber": "^3.1.3",
     "turndown": "^5.0.1",
     "unzalgo": "^2.1.2",
@@ -61,8 +61,8 @@
   },
   "devDependencies": {
     "@aquarius/testing": "^0.0.1",
-    "@babel/cli": "^7.8.3",
-    "@babel/core": "^7.8.3",
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.8.4",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.3",
@@ -73,8 +73,8 @@
     "babel-preset-env": "^1.6.1",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-config-prettier": "^6.9.0",
-    "eslint-plugin-import": "^2.20.0",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jest": "^23.6.0",
     "eslint-plugin-prettier": "^3.1.2",
     "jest": "^25.1.0",
